requirements.freeze.txt
addict==2.4.0
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
defusedxml==0.7.1
fast_ctc_decode==0.3.2
fastapi==0.78.0
fastjsonschema==2.15.3
filelock==3.7.1
grpcio==1.46.3
h11==0.13.0
huggingface-hub==0.7.0
idna==3.3
imagecodecs==2022.2.22
imageio==2.19.3
iniconfig==1.1.1
joblib==1.1.0
jstyleson==0.0.2
lmdb==1.3.0
mlserver==1.0.1
networkx==2.8.3
nibabel==3.2.2
nltk==3.7
numpy==1.19.5
onnx==1.11.0
opencv-python==4.5.5.64
openvino==2022.1.0
openvino-dev==2022.1.0
openvino-telemetry==2022.1.1
packaging==21.3
pandas==1.1.5
parasail==1.2.4
Pillow==9.1.1
pluggy==1.0.0
progress==1.6
prometheus-client==0.14.1
protobuf==4.21.1
py==1.11.0
py-cpuinfo==8.0.0
py-grpc-prometheus==0.7.0
pyclipper==1.3.0.post3
pydantic==1.9.1
pydicom==2.3.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
rawpy==0.17.1
regex==2022.6.2
requests==2.28.0
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.96
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
starlette-exporter==0.13.0
texttable==1.6.4
threadpoolctl==3.1.0
tifffile==2022.5.4
tokenizers==0.10.3
tomli==2.0.1
tqdm==4.64.0
transformers==4.16.2
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
